jackson auto truncate float to integer during serialization (ACCEPT_FLOAT_AS_INT flag). This causes volume value truncate to integer (eg: from 1.5 to 1) and not throwing validation error.

+++ api/src/main/java/uk/gov/hmcts/payment/api/dto/mapper/CardPaymentDtoMapper.java
@@ -103,7 +103,7 @@ public class CardPaymentDtoMapper {
             .calculatedAmount(feeDto.getCalculatedAmount())
             .code(feeDto.getCode())
             .version(feeDto.getVersion())
-            .volume(feeDto.getVolume() == null ? 1 : feeDto.getVolume())
+            .volume(feeDto.getVolume() == null ? 1 : feeDto.getVolume().intValue())
             .build();
     }
 
@@ -112,7 +112,7 @@ public class CardPaymentDtoMapper {
             .calculatedAmount(fee.getCalculatedAmount())
             .code(fee.getCode())
             .version(fee.getVersion())
-            .volume(fee.getVolume())
+            .volume(fee.getVolume().doubleValue())
             .build();
 
     }
